requirements.txt
bidict==0.21.4
click==8.0.3
Flask==2.0.2
Flask-Login==0.5.0
Flask-SocketIO==5.1.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
python-decouple==3.5
python-engineio==4.3.0
python-socketio==5.4.1
SQLAlchemy==1.4.26
Werkzeug==2.0.2
